+++ Assets/Editor/Scripts/ScriptableObjects/PedeSettings.cs
@@ -1,12 +1,17 @@
+using Thisaislan.PersistenceEasyToDeleteInEditor.Editor.Constants;
 using Thisaislan.PersistenceEasyToDeleteInEditor.Editor.ScriptableObjects.Data;
 using UnityEngine;
 
 namespace Thisaislan.PersistenceEasyToDeleteInEditor.Editor.ScriptableObjects.Settings
 {
-    /// <summary> ScriptableObject containing settings for the Pede works. </summary>
-    public class PedeSettings : ScriptableObject
+    internal class PedeSettings : ScriptableObject
     {
-        /// <summary> Data used in editor. </summary>
-        public PedeData pedeData;
+        
+        [Header(Metadata.PedeSettingsHeaderAttr)]
+        [Tooltip(Metadata.PedeSettingsTooltipAttr)]
+        [Space]
+        [SerializeField]
+        internal PedeData pedeData;
+        
     }
 }