+++ Assets/Runtime/Scripts/Consts.cs
@@ -4,10 +4,12 @@ namespace Thisaislan.PersistenceEasyToDeleteInEditor.Constants
 {
     internal static class Consts
     {
+        
         internal const string PedeKeyFormat = "{0}~{1}";
         
         internal const string PedeFilePethFormat = "{0}/{1}";
         
         internal static readonly string PedeFileRootFolderName = $"{Application.persistentDataPath}/Pede";
+        
     }
 }